Fixed free memory used problem and got rid of warnings
This problem was recently introduced with the client API timeout changes

+++ src/cc/Hypertable/Lib/TableMutatorScatterBuffer.h
@@ -92,7 +92,7 @@ namespace Hypertable {
 	boost::xtime expire_time;
 
 	boost::xtime_get(&expire_time, boost::TIME_UTC);
-	expire_time.sec += timer.remaining();
+	expire_time.sec += (int64_t)timer.remaining();
 
 	while (m_outstanding) {
 	  if (!m_cond.timed_wait(lock, expire_time))
